+++ xenon-common/src/test/java/com/vmware/xenon/common/TestServiceHost.java
@@ -841,6 +841,7 @@ public class TestServiceHost {
 
         long cacheMissCount = 0;
         long cacheClearCount = 0;
+        ServiceStat cacheClearStat = null;
         Map<URI, ServiceStats> servicesWithMaintenance = new HashMap<>();
 
         // guarantee at least a few intervals have passed. Other we risk false negatives.
@@ -904,9 +905,21 @@ public class TestServiceHost {
                 continue;
             }
 
+            Map<String, ServiceStat> mgmtStats = getManagementServiceStats();
+            cacheClearStat = mgmtStats.get(Service.STAT_NAME_SERVICE_CACHE_CLEAR_COUNT);
+            if (cacheClearStat == null || cacheClearStat.latestValue < 1) {
+                this.host.log("Cache clear stat on management service not seen");
+                Thread.sleep(maintIntervalMillis * 2);
+                continue;
+            }
             break;
         }
         long end = Utils.getNowMicrosUtc();
+
+        if (cacheClearStat == null || cacheClearStat.latestValue < 1) {
+            throw new IllegalStateException(
+                    "Cache clear stat on management service not observed");
+        }
 
         this.host.log("State cache misses: %d, cache clears: %d", cacheMissCount, cacheClearCount);
 
@@ -1387,18 +1400,26 @@ public class TestServiceHost {
         this.host.testWait();
 
         if (this.testDurationSeconds == 0) {
-            for (ServiceStats statsPerInstance : stats.values()) {
-                ServiceStat pauseStat = statsPerInstance.entries.get(Service.STAT_NAME_PAUSE_COUNT);
-                ServiceStat resumeStat = statsPerInstance.entries
-                        .get(Service.STAT_NAME_RESUME_COUNT);
-                if (pauseStat == null) {
-                    throw new IllegalStateException("No pauses observed");
+            this.host.waitFor("Service stats did not get updated", () -> {
+                for (ServiceStats statsPerInstance : stats.values()) {
+                    ServiceStat pauseStat = statsPerInstance.entries.get(Service.STAT_NAME_PAUSE_COUNT);
+                    ServiceStat resumeStat = statsPerInstance.entries.get(Service.STAT_NAME_RESUME_COUNT);
+                    if (pauseStat == null || resumeStat == null) {
+                        return false;
+                    }
                 }
-                if (resumeStat == null) {
-                    throw new IllegalStateException("No resumes observed");
+
+                Map<String, ServiceStat> mgmtStats = getManagementServiceStats();
+                ServiceStat mgmtPauseStat = mgmtStats.get(Service.STAT_NAME_SERVICE_PAUSE_COUNT);
+                ServiceStat mgmtResumeStat = mgmtStats.get(Service.STAT_NAME_SERVICE_RESUME_COUNT);
+                if (mgmtPauseStat == null || mgmtResumeStat == null ||
+                        (int)mgmtPauseStat.latestValue < states.size() ||
+                        (int)mgmtPauseStat.latestValue < states.size()) {
+                    return false;
                 }
-            }
-            return;
+
+                return true;
+            });
         }
 
         states.clear();
@@ -1465,7 +1486,7 @@ public class TestServiceHost {
     }
 
     @Test
-    public void serviceStopDueToMemoryPressure() throws Throwable {
+    public void maintenanceForOnDemandLoadServices() throws Throwable {
         setUp(true);
 
         long maintenanceIntervalMillis = 100;
@@ -1473,8 +1494,6 @@ public class TestServiceHost {
                 .toMicros(maintenanceIntervalMillis);
 
         // induce host to clear service state cache by setting mem limit low
-        this.host.setServiceMemoryLimit(ServiceHost.ROOT_PATH, 0.0001);
-        this.host.setServiceMemoryLimit(LuceneDocumentIndexService.SELF_LINK, 0.0001);
         this.host.setMaintenanceIntervalMicros(maintenanceIntervalMicros);
         this.host.setServiceCacheClearDelayMicros(maintenanceIntervalMicros / 2);
         this.host.start();
@@ -1493,9 +1512,6 @@ public class TestServiceHost {
         // guarantee at least a few intervals have passed.
         Thread.sleep(maintenanceIntervalMillis * 10);
 
-        // restore memory limit so we don't keep re-pausing/resuming services every time we talk to them.
-        this.host.setServiceMemoryLimit(ServiceHost.ROOT_PATH, ServiceHost.DEFAULT_PCT_MEMORY_LIMIT);
-
         // Let's verify now that all of the services have stopped by now.
         int stoppedCount = 0;
         Date exp = this.host.getTestExpiration();
@@ -1507,14 +1523,31 @@ public class TestServiceHost {
                     stoppedCount++;
                 }
             }
-
-            if (stoppedCount == this.serviceCount) {
-                break;
+            if (stoppedCount < this.serviceCount) {
+                Thread.sleep(maintenanceIntervalMillis / 2);
+                continue;
             }
+
+            Map<String, ServiceStat> stats = getManagementServiceStats();
+            ServiceStat odlStops = stats.get(Service.STAT_NAME_ODL_STOP_COUNT);
+            ServiceStat odlCacheClears = stats.get(Service.STAT_NAME_ODL_CACHE_CLEAR_COUNT);
+            ServiceStat cacheClears = stats.get(Service.STAT_NAME_SERVICE_CACHE_CLEAR_COUNT);
+            if (odlStops == null || odlStops.latestValue != this.serviceCount) {
+                Thread.sleep(maintenanceIntervalMillis / 2);
+                continue;
+            }
+            if (odlCacheClears == null || odlStops.latestValue != this.serviceCount) {
+                Thread.sleep(maintenanceIntervalMillis / 2);
+                continue;
+            }
+            if (cacheClears == null || cacheClears.latestValue != this.serviceCount) {
+                Thread.sleep(maintenanceIntervalMillis / 2);
+                continue;
+            }
+            return;
         }
 
-        this.host.log("Total services stopped:" + stoppedCount);
-        assertTrue("Not all services stopped.", stoppedCount == this.serviceCount);
+        throw new TimeoutException();
     }
 
     private void patchExampleServices(Map<URI, ExampleServiceState> states, int count)
@@ -1531,6 +1564,25 @@ public class TestServiceHost {
             }
         }
         this.host.testWait();
+    }
+
+    private Map<String, ServiceStat> getManagementServiceStats() throws Throwable {
+        Map<String, ServiceStat> stats = Collections.synchronizedMap(new HashMap<>());
+        this.host.testStart(1);
+        Operation getManagementStats = Operation.createGet(
+                UriUtils.buildStatsUri(this.host.getManagementServiceUri()))
+                .setCompletion((o, e) -> {
+                    if (e != null) {
+                        this.host.failIteration(e);
+                        return;
+                    }
+                    ServiceStats response = o.getBody(ServiceStats.class);
+                    stats.putAll(response.entries);
+                    this.host.completeIteration();
+                });
+        this.host.send(getManagementStats);
+        this.host.testWait();
+        return stats;
     }
 
     @Test